Completed migrations and test

+++ tests/utils.py
@@ -6,12 +6,18 @@ from blog.constants import BLOG_REDIS_HOST, BLOG_REDIS_PORT
 from blog.db import client
 
 
-def drop_database():
-    """Drop all collections in database"""
+def drop_database(drop_redis: bool=True):
+    """
+    Drop all collections in database.
+
+    :param drop_redis: If toggled, will flush redis database.
+    :type drop_redis: bool
+    """
     db_name = client.get_database().name
     client.drop_database(db_name)
-    redis = StrictRedis(host=BLOG_REDIS_HOST, port=BLOG_REDIS_PORT)
-    redis.flushdb()
+    if drop_redis:
+        redis = StrictRedis(host=BLOG_REDIS_HOST, port=BLOG_REDIS_PORT)
+        redis.flushdb()
 
 
 def random_string(length: int) -> str:
